Delete and update recipe authorisation

+++ controllers/recipesControllers.js
@@ -3,6 +3,7 @@ import jwt from "jsonwebtoken";
 import "dotenv/config.js";
 import { User } from "../models/users.js";
 import Joi from "joi";
+import { response } from "express";
 
 //Get all recipes
 export const allRecipes = async (req, res) => {
@@ -13,7 +14,7 @@ export const allRecipes = async (req, res) => {
   res.send(recipes);
 };
 
-//Get favorite recipes
+//Get favorite recipes array
 export const favoriteRecipes = async (req, res) => {
   jwt.verify(
     req.cookies.jwt,
@@ -34,7 +35,7 @@ export const favoriteRecipes = async (req, res) => {
   );
 };
 
-//Get favorites array
+//Get favorite recipes ids array
 export const favoritesArray = async (req, res) => {
   jwt.verify(
     req.cookies.jwt,
@@ -70,18 +71,43 @@ export const newRecipe = async (req, res) => {
 
 //Delete a recipe
 export const deleteRecipe = async (req, res) => {
-  await Recipe.findByIdAndDelete(req.params.id)
-    .catch((err) => res.status(404).send(err))
-    .then(() => res.send("Recipe deleted"));
+  jwt.verify(
+    req.cookies.jwt,
+    process.env.ACCESS_TOKEN_SECRET,
+    async (error, decodedToken) => {
+      if (error) return res.status(404).send(error.message);
+      const recipe = await Recipe.findById(req.params.id).catch((err) =>
+        response.status(404).send(err)
+      );
+      //allow only the author to delete the recipe
+      if (decodedToken.name === recipe.author)
+        await Recipe.findByIdAndDelete(req.params.id)
+          .catch((err) => res.status(404).send(err))
+          .then(() => res.send("Recipe deleted"));
+      else res.status(403).send("You aren't allowed to delete other's recipes");
+    }
+  );
 };
 
 //Edit recipe
 export const editRecipe = async (req, res) => {
-  const { error } = recipeSchema.validate(req.body);
-  if (error) return res.status(400).send(error.details[0].message);
-  await Recipe.findByIdAndUpdate(req.params.id, { ...req.body })
-    .catch((err) => res.status(404).send(err))
-    .then(() => res.send("Recipe updated"));
+  jwt.verify(
+    req.cookies.jwt,
+    process.env.ACCESS_TOKEN_SECRET,
+    async (error, decodedToken) => {
+      if (error) return res.status(404).send(error.message);
+
+      //allow only the author to edit the recipe
+      if (decodedToken.name === req.body.author) {
+        const { error } = recipeSchema.validate(req.body);
+        if (error) return res.status(400).send(error.details[0].message);
+        await Recipe.findByIdAndUpdate(req.params.id, { ...req.body })
+          .catch((err) => res.status(404).send(err))
+          .then(() => res.send("Recipe updated"));
+      } else
+        res.status(403).send("You are not allowed to edit other's recipes");
+    }
+  );
 };
 
 //Like or dislike recipe
